requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Caching
Werkzeug
Flask-Login
Flask-Bcrypt
openpyxl
gunicorn
